requirements.txt
google-generativeai
yt-dlp
